rename encoder_pt to encoder
