+++ sockeye/train.py
@@ -35,7 +35,7 @@ from . import arguments
 from . import checkpoint_decoder
 from . import constants as C
 from . import data_io
-from . import encoder_pt
+from . import encoder
 from . import layers_pt
 from . import loss_pt
 from . import lr_scheduler
