+++ test/unit/test_params.py
@@ -20,7 +20,7 @@ import pytest
 import torch as pt
 
 import sockeye.constants as C
-import sockeye.encoder_pt
+import sockeye.encoder
 import sockeye.model_pt
 import sockeye.training
